# src/bb_paxdata/interfaces/web/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BB-PAXDATA API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/queue": {
      "get": {
        "summary": "Get Fail Queue",
        "description": "Retrieve failed validation logs with sentence details and AI scores, ordered by priority score.",
        "parameters": [
          {
            "name": "formula_name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FailQueueItemResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/queue/context/{sent_id}": {
      "get": {
        "summary": "Get Triplet Context",
        "description": "Retrieve the target sentence, its previous and next sentence, and speaker/panel metadata.",
        "parameters": [
          {
            "name": "sent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripletContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/queue/similar/{sent_id}/{formula_name}": {
      "get": {
        "summary": "Get Similar Cases",
        "description": "Retrieve similar historical failures for the same validation formula to guide the reviewer.",
        "parameters": [
          {
            "name": "sent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formula_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimilarCaseResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FailQueueItemResponse": {
        "type": "object",
        "properties": {
          "log_id": {
            "type": "integer"
          },
          "run_id": {
            "type": "string"
          },
          "formula_name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "actual_value": {
            "type": "number"
          },
          "expected_constraint": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "human_verdict": {
            "type": "string",
            "nullable": true
          },
          "human_note": {
            "type": "string",
            "nullable": true
          },
          "log_version": {
            "type": "integer"
          },
          "reviewer_id": {
            "type": "string",
            "nullable": true
          },
          "sentence_text": {
            "type": "string",
            "nullable": true
          },
          "sent_id": {
            "type": "string",
            "nullable": true
          },
          "seg_id": {
            "type": "string",
            "nullable": true
          },
          "file_id": {
            "type": "string",
            "nullable": true
          },
          "speaker_name": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "power_level": {
            "type": "number"
          },
          "ai_risk_score": {
            "type": "number",
            "nullable": true
          },
          "ai_emotion_category": {
            "type": "string",
            "nullable": true
          },
          "ai_diplomatic_tone": {
            "type": "string",
            "nullable": true
          },
          "review_id": {
            "type": "integer",
            "nullable": true
          },
          "review_status": {
            "type": "string",
            "nullable": true
          },
          "trigger_type": {
            "type": "string",
            "nullable": true
          },
          "priority_score": {
            "type": "number"
          },
          "created_at": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "PanelContext": {
        "type": "object",
        "properties": {
          "file_id": {
            "type": "string"
          },
          "panel_number": {
            "type": "integer",
            "nullable": true
          },
          "date": {
            "type": "string",
            "nullable": true
          },
          "theme": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SimilarCaseResponse": {
        "type": "object",
        "properties": {
          "log_id": {
            "type": "integer"
          },
          "entity_id": {
            "type": "string"
          },
          "actual_value": {
            "type": "number"
          },
          "human_verdict": {
            "type": "string",
            "nullable": true
          },
          "human_note": {
            "type": "string",
            "nullable": true
          },
          "sentence_text": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "speaker_name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SpeakerContext": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "power_level": {
            "type": "number"
          },
          "influence_tier": {
            "type": "string",
            "nullable": true
          },
          "bloc": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TripletContextResponse": {
        "type": "object",
        "properties": {
          "prev": {
            "type": "string",
            "nullable": true
          },
          "current": {
            "type": "string",
            "nullable": true
          },
          "next": {
            "type": "string",
            "nullable": true
          },
          "speaker": {
            "$ref": "#/components/schemas/SpeakerContext",
            "nullable": true
          },
          "panel": {
            "$ref": "#/components/schemas/PanelContext",
            "nullable": true
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "loc": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "input": {
            "nullable": true
          },
          "ctx": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}